♻️ refactor : setup-textarea-styles로 분리  #30
